Zero released memory pages
When a task ends and its memory pages are released they are now zero filled. This is particulary important for pages containing entries for the various page tables; otherwise they might be interpreted as valid table entries. For speed this is done with a small assembler routine in tasking.s.